requirements.txt
Django>=6.0,<6.1
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-jazzmin
django-import-export
python-decouple
psycopg2-binary
Pillow
stripe
